refactor: relocate refresh-token cleanup scheduler into auth
- Relocate SchedulerConfig (runtime cleanup task) to features/auth/infra as RefreshTokensCleanupScheduler,
  so refresh-token cleanup lives with the auth vertical slice it serves; drop the stale TODO/move marker
- Flatten server/config/modules/{startup,runtime} into a single modules package
- Fold the ServerEnvironment interface into a nested ServerConfig.Env
