+++ backend/api/src/main/java/com/anibalxyz/features/auth/infra/RefreshTokensCleanupScheduler.java
@@ -1,9 +1,9 @@
-package com.anibalxyz.server.config.modules.runtime;
+package com.anibalxyz.features.auth.infra;
 
 import static net.logstash.logback.argument.StructuredArguments.kv;
 
 import com.anibalxyz.features.auth.application.RefreshTokenService;
-import com.anibalxyz.server.config.modules.startup.StartupConfig;
+import com.anibalxyz.server.config.modules.StartupConfig;
 import io.javalin.config.JavalinConfig;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
@@ -11,24 +11,16 @@ import java.util.concurrent.TimeUnit;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-// TODO: move to correspondent vertical slice
-/**
- * Manages background tasks and periodic maintenance.
- *
- * <p>Initializes a {@link ScheduledExecutorService} for recurring tasks and ensures a graceful
- * shutdown when the server stops.
- */
-public class SchedulerConfig implements StartupConfig {
+public class RefreshTokensCleanupScheduler implements StartupConfig {
 
-  private static final Logger log = LoggerFactory.getLogger(SchedulerConfig.class);
+  private static final Logger log = LoggerFactory.getLogger(RefreshTokensCleanupScheduler.class);
   private final RefreshTokenService refreshTokenService;
   private ScheduledExecutorService scheduler;
 
-  public SchedulerConfig(RefreshTokenService refreshTokenService) {
+  public RefreshTokensCleanupScheduler(RefreshTokenService refreshTokenService) {
     this.refreshTokenService = refreshTokenService;
   }
 
-  /** Initializes schedules and registers shutdown hooks. */
   @Override
   public void apply(JavalinConfig cfg) {
     scheduler = Executors.newSingleThreadScheduledExecutor();
